+++ blockindex/liquidstaking_indexer_test.go
@@ -1,0 +1,12 @@
+// Copyright (c) 2023 IoTeX Foundation
+// This source code is provided 'as is' and no warranties are given as to title or non-infringement, merchantability
+// or fitness for purpose and, to the extent permitted by law, all liability for your use of the code is disclaimed.
+// This source code is governed by Apache License 2.0 that can be found in the LICENSE file.
+
+package blockindex
+
+import "testing"
+
+func TestLiquidStakingIndexer(t *testing.T) {
+
+}
