+++ e2etest/liquid_staking_test.go
@@ -1,0 +1,161 @@
+package e2etest
+
+import (
+	"context"
+	"encoding/hex"
+	"math/big"
+	"testing"
+	"time"
+
+	"github.com/iotexproject/go-pkgs/hash"
+	"github.com/iotexproject/iotex-core/action"
+	"github.com/iotexproject/iotex-core/action/protocol"
+	"github.com/iotexproject/iotex-core/blockchain/genesis"
+	"github.com/iotexproject/iotex-core/config"
+	"github.com/iotexproject/iotex-core/pkg/unit"
+	"github.com/iotexproject/iotex-core/server/itx"
+	"github.com/iotexproject/iotex-core/state"
+	"github.com/iotexproject/iotex-core/test/identityset"
+	"github.com/iotexproject/iotex-core/testutil"
+	"github.com/stretchr/testify/require"
+)
+
+const (
+	_liquidStakingContractByteCode = `60806040523480156200001157600080fd5b506040518060400160405280600981526020017f4275636b65744e465400000000000000000000000000000000000000000000008152506040518060400160405280600381526020017f424b54000000000000000000000000000000000000000000000000000000000081525081600090816200008f91906200042d565b508060019081620000a191906200042d565b505050620000c4620000b8620000e560201b60201c565b620000ed60201b60201c565b6000600660146101000a81548160ff02191690831515021790555062000514565b600033905090565b6000600660009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905081600660006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055508173ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff167f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e060405160405180910390a35050565b600081519050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b600060028204905060018216806200023557607f821691505b6020821081036200024b576200024a620001ed565b5b50919050565b60008190508160005260206000209050919050565b60006020601f8301049050919050565b600082821b905092915050565b600060088302620002b57fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff8262000276565b620002c1868362000276565b95508019841693508086168417925050509392505050565b6000819050919050565b6000819050919050565b60006200030e620003086200030284620002d9565b620002e3565b620002d9565b9050919050565b6000819050919050565b6200032a83620002ed565b62000342620003398262000315565b84845462000283565b825550505050565b600090565b620003596200034a565b620003668184846200031f565b505050565b5b818110156200038e57620003826000826200034f565b6001810190506200036c565b5050565b601f821115620003dd57620003a78162000251565b620003b28462000266565b81016020851015620003c2578190505b620003da620003d18562000266565b8301826200036b565b50505b505050565b600082821c905092915050565b60006200040260001984600802620003e2565b1980831691505092915050565b60006200041d8383620003ef565b9150826002028217905092915050565b6200043882620001b3565b67ffffffffffffffff811115620004545762000453620001be565b5b6200046082546200021c565b6200046d82828562000392565b600060209050601f831160018114620004a5576000841562000490578287015190505b6200049c85826200040f565b8655506200050c565b601f198416620004b58662000251565b60005b82811015620004df57848901518255600182019150602085019450602081019050620004b8565b86831015620004ff5784890151620004fb601f891682620003ef565b8355505b6001600288020188555050505b505050505050565b61700c80620005246000396000f3fe6080604052600436106102925760003560e01c806378bfca101161015a578063c87b56dd116100c1578063e0028ecf1161007a578063e0028ecf14610a16578063e449f34114610a3f578063e985e9c514610a68578063eb0ffb2e14610aa5578063f0b56b5d14610ace578063f2fde38b14610af957610292565b8063c87b56dd146108fc578063c8e7792314610939578063cd0a02d014610962578063d0949f9914610992578063d6605fd8146109bd578063d6819bcc146109e657610292565b8063a22cb46511610113578063a22cb46514610820578063ad46fc6414610849578063b2383e5514610872578063b88d4fde1461088e578063b8f4bd7b146108b7578063bbe33ea5146108e057610292565b806378bfca101461070e5780638456cb591461074b5780638da5cb5b1461076257806393b6ef591461078d57806395d89b41146107ca5780639f7d5b00146107f557610292565b806342842e0e116101fe5780635d36598f116101b75780635d36598f146105ee5780636198e339146106175780636352211e146106405780636faa5c271461067d57806370a08231146106ba578063715018a6146106f757610292565b806342842e0e146104c3578063431cd92a146104ec57806343e06c591461052d578063597cc14a1461056a5780635c975abb1461059a5780635ceb8b5b146105c557610292565b80631338736f116102505780631338736f146103cb57806323b872dd146103f45780632dc830081461041d5780632e17de78146104465780633f4ba83a1461046f5780633fac69af1461048657610292565b8062f714ce1461029757806301ffc9a7146102c057806303459b16146102fd57806306fdde031461033a578063081812fc14610365578063095ea7b3146103a2575b600080fd5b3480156102a357600080fd5b506102be60048036038101906102b99190614f10565b610b22565b005b3480156102cc57600080fd5b506102e760048036038101906102e29190614fa8565b610bf7565b6040516102f49190614ff0565b60405180910390f35b34801561030957600080fd5b50610324600480360381019061031f919061500b565b610cd9565b6040516103319190615047565b60405180910390f35b34801561034657600080fd5b5061034f610d0a565b60405161035c91906150f2565b60405180910390f35b34801561037157600080fd5b5061038c6004803603810190610387919061500b565b610d9c565b6040516103999190615135565b60405180910390f35b3480156103ae57600080fd5b506103c960048036038101906103c4919061517c565b610de2565b005b3480156103d757600080fd5b506103f260048036038101906103ed91906151bc565b610ef9565b005b34801561040057600080fd5b5061041b600480360381019061041691906151fc565b610f73565b005b34801561042957600080fd5b50610444600480360381019061043f91906152a7565b610fd3565b005b34801561045257600080fd5b5061046d6004803603810190610468919061500b565b61103e565b005b34801561047b57600080fd5b506104846110f6565b005b34801561049257600080fd5b506104ad60048036038101906104a8919061534c565b611108565b6040516104ba9190615519565b60405180910390f35b3480156104cf57600080fd5b506104ea60048036038101906104e591906151fc565b6112cb565b005b3480156104f857600080fd5b50610513600480360381019061050e919061500b565b6112eb565b60405161052495949392919061554a565b60405180910390f35b34801561053957600080fd5b50610554600480360381019061054f91906151bc565b611378565b6040516105619190614ff0565b60405180910390f35b610584600480360381019061057f91906152a7565b611394565b6040516105919190615047565b60405180910390f35b3480156105a657600080fd5b506105af611413565b6040516105bc9190614ff0565b60405180910390f35b3480156105d157600080fd5b506105ec60048036038101906105e791906155f3565b61142a565b005b3480156105fa57600080fd5b5061061560048036038101906106109190615653565b6114e0565b005b34801561062357600080fd5b5061063e6004803603810190610639919061500b565b611589565b005b34801561064c57600080fd5b506106676004803603810190610662919061500b565b6115f6565b6040516106749190615135565b60405180910390f35b34801561068957600080fd5b506106a4600480360381019061069f919061534c565b61167c565b6040516106b19190615519565b60405180910390f35b3480156106c657600080fd5b506106e160048036038101906106dc91906156a0565b61183f565b6040516106ee9190615047565b60405180910390f35b34801561070357600080fd5b5061070c6118f6565b005b34801561071a57600080fd5b50610735600480360381019061073091906151bc565b61190a565b60405161074291906157be565b60405180910390f35b34801561075757600080fd5b50610760611a50565b005b34801561076e57600080fd5b50610777611a62565b6040516107849190615135565b60405180910390f35b34801561079957600080fd5b506107b460048036038101906107af919061500b565b611a8c565b6040516107c19190615047565b60405180910390f35b3480156107d657600080fd5b506107df611ac8565b6040516107ec91906150f2565b60405180910390f35b34801561080157600080fd5b5061080a611b5a565b6040516108179190615047565b60405180910390f35b34801561082c57600080fd5b506108476004803603810190610842919061580c565b611b67565b005b34801561085557600080fd5b50610870600480360381019061086b919061584c565b611b7d565b005b61088c600480360381019061088791906151bc565b611c28565b005b34801561089a57600080fd5b506108b560048036038101906108b091906159dc565b611dfc565b005b3480156108c357600080fd5b506108de60048036038101906108d99190615a5f565b611e5e565b005b6108fa60048036038101906108f591906155f3565b611f6f565b005b34801561090857600080fd5b50610923600480360381019061091e919061500b565b6122e9565b60405161093091906150f2565b60405180910390f35b34801561094557600080fd5b50610960600480360381019061095b91906151bc565b612351565b005b61097c60048036038101906109779190615abf565b6124d4565b6040516109899190615047565b60405180910390f35b34801561099e57600080fd5b506109a76125cc565b6040516109b49190615047565b60405180910390f35b3480156109c957600080fd5b506109e460048036038101906109df91906151bc565b6125f0565b005b610a0060048036038101906109fb9190615be9565b6127b9565b604051610a0d9190615047565b60405180910390f35b348015610a2257600080fd5b50610a3d6004803603810190610a3891906151bc565b6128da565b005b348015610a4b57600080fd5b50610a666004803603810190610a619190615653565b612952565b005b348015610a7457600080fd5b50610a8f6004803603810190610a8a9190615c58565b612a46565b604051610a9c9190614ff0565b60405180910390f35b348015610ab157600080fd5b50610acc6004803603810190610ac791906151bc565b612ada565b005b348015610ada57600080fd5b50610ae3612b72565b604051610af09190615047565b60405180910390f35b348015610b0557600080fd5b50610b206004803603810190610b1b91906156a0565b612b78565b005b610b2a612bfb565b81610b3481612c45565b60006008600085815260200190815260200160002090506000610b5a8260020154612cbe565b14610b9a576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610b9190615ce4565b60405180910390fd5b610ba384612d37565b610bad8184612e85565b8273ffffffffffffffffffffffffffffffffffffffff16847fd964a27d45f595739c13d8b1160b57491050cacf3a2e5602207277d6228f64ee60405160405180910390a350505050565b60007f80ac58cd000000000000000000000000000000000000000000000000000000007bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916827bffffffffffffffffffffffffffffffffffffffffffffffffffffffff19161480610cc257507f5b5e139f000000000000000000000000000000000000000000000000000000007bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916827bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916145b80610cd25750610cd182612f65565b5b9050919050565b6000610ce482612fcf565b610d036008600084815260200190815260200160002060020154612cbe565b9050919050565b606060008054610d1990615d33565b80601f0160208091040260200160405190810160405280929190818152602001828054610d4590615d33565b8015610d925780601f10610d6757610100808354040283529160200191610d92565b820191906000526020600020905b815481529060010190602001808311610d7557829003601f168201915b5050505050905090565b6000610da78261301a565b6004600083815260200190815260200160002060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff169050919050565b6000610ded826115f6565b90508073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff1603610e5d576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610e5490615dd6565b60405180910390fd5b8073ffffffffffffffffffffffffffffffffffffffff16610e7c613065565b73ffffffffffffffffffffffffffffffffffffffff161480610eab5750610eaa81610ea5613065565b612a46565b5b610eea576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610ee190615e68565b60405180910390fd5b610ef4838361306d565b505050565b610f01612bfb565b81610f0b81612c45565b6000600860008581526020019081526020016000209050610f2b81613126565b610f358184613176565b837f907fece23ce39fbcbceb71e515043fe29408353fbb393b25b35eb8a70a4bad0b84604051610f659190615047565b60405180910390a250505050565b610f84610f7e613065565b826133b7565b610fc3576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610fba90615efa565b60405180910390fd5b610fce83838361344c565b505050565b610fdb612bfb565b81610fe581612c45565b6110016008600085815260200190815260200160002083613745565b827ffec7db38481afeb8686a62ee7bba420143bd43540fe5e57b7316be50bdaa220c836040516110319190615f1a565b60405180910390a2505050565b611046612bfb565b8061105081612c45565b600060086000848152602001908152602001600020905061107081613126565b600061107b82613ae7565b146110bb576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016110b290615f81565b60405180910390fd5b6110c481613b91565b827f11725367022c3ff288940f4b5473aa61c2da6a24af7363a1128ee2401e8983b260405160405180910390a2505050565b6110fe613c78565b611106613cf6565b565b60608282905067ffffffffffffffff811115611127576111266158b1565b5b60405190808252806020026020018201604052801561115a57816020015b60608152602001906001900390816111455790505b5090506000611167611b5a565b905060005b848490508110156112c3578167ffffffffffffffff811115611191576111906158b1565b5b6040519080825280602002602001820160405280156111bf5781602001602082028036833780820191505090505b508382815181106111d3576111d2615fa1565b5b60200260200101819052506000600960008787858181106111f7576111f6615fa1565b5b905060200201602081019061120c9190615fd0565b73ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020905060005b838110156112b1578160008281526020019081526020016000205485848151811061127b5761127a615fa1565b5b6020026020010151828151811061129557611294615fa1565b5b6020026020010181815250506112aa81613d59565b905061124d565b50506112bc81613d59565b905061116c565b505092915050565b6112e683838360405180602001604052806000815250611dfc565b505050565b60008060008060006112fc86612fcf565b60006008600088815260200190815260200160002090506000600b82600001548154811061132d5761132c615fa1565b5b9060005260206000209060030201905080600001548160010154836001015484600201548560030160009054906101000a900460a01b96509650965096509650505091939590929450565b600061138c6113878484613d66565b613de5565b905092915050565b600061139e612bfb565b600034905060006113af8286613d66565b90506113ba81613e17565b6113c48185613e62565b60006007549050807f1f44b78b04f7c6f80fc97ae8b196d9e9d7a81663114744f18fe5d073cd70ce4a8685896040516113ff93929190615ffd565b60405180910390a280935050505092915050565b6000600660149054906101000a900460ff16905090565b611432612bfb565b60008060005b858590508110156114d85785858281811061145657611455615fa1565b5b90506020020135925061146883612c45565b60086000848152602001908152602001600020915061148682613126565b6114908285613176565b827f907fece23ce39fbcbceb71e515043fe29408353fbb393b25b35eb8a70a4bad0b856040516114c09190615047565b60405180910390a26114d181613d59565b9050611438565b505050505050565b6114e8612bfb565b60008060005b848490508110156115825784848281811061150c5761150b615fa1565b5b90506020020135925061151e83612c45565b60086000848152602001908152602001600020915061153c8261400e565b6115458261405e565b827ff27b6ce5b2f5e68ddb2fd95a8a909d4ecf1daaac270935fff052feacb24f184260405160405180910390a261157b81613d59565b90506114ee565b5050505050565b611591612bfb565b8061159b81612c45565b60006008600084815260200190815260200160002090506115bb8161400e565b6115c48161405e565b827ff27b6ce5b2f5e68ddb2fd95a8a909d4ecf1daaac270935fff052feacb24f184260405160405180910390a2505050565b600080611602836141ed565b9050600073ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff1603611673576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161166a90616080565b60405180910390fd5b80915050919050565b60608282905067ffffffffffffffff81111561169b5761169a6158b1565b5b6040519080825280602002602001820160405280156116ce57816020015b60608152602001906001900390816116b95790505b50905060006116db611b5a565b905060005b84849050811015611837578167ffffffffffffffff811115611705576117046158b1565b5b6040519080825280602002602001820160405280156117335781602001602082028036833780820191505090505b5083828151811061174757611746615fa1565b5b60200260200101819052506000600a600087878581811061176b5761176a615fa1565b5b90506020020160208101906117809190615fd0565b73ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020905060005b8381101561182557816000828152602001908152602001600020548584815181106117ef576117ee615fa1565b5b6020026020010151828151811061180957611808615fa1565b5b60200260200101818152505061181e81613d59565b90506117c1565b505061183081613d59565b90506116e0565b505092915050565b60008073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff16036118af576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016118a690616112565b60405180910390fd5b600360008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020549050919050565b6118fe613c78565b611908600061422a565b565b606060008211801561192e575061191f611b5a565b828461192b9190616161565b11155b61196d576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611964906161e1565b60405180910390fd5b8167ffffffffffffffff811115611987576119866158b1565b5b6040519080825280602002602001820160405280156119c057816020015b6119ad614e47565b8152602001906001900390816119a55790505b50905060005b82811015611a4957600b818501815481106119e4576119e3615fa1565b5b90600052602060002090600302016040518060600160405290816000820154815260200160018201548152602001600282015481525050828281518110611a2e57611a2d615fa1565b5b6020026020010181905250611a4281613d59565b90506119c6565b5092915050565b611a58613c78565b611a606142f0565b565b6000600660009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905090565b6000611a9782612fcf565b6000600860008481526020019081526020016000209050611ab781613126565b611ac081613ae7565b915050919050565b606060018054611ad790615d33565b80601f0160208091040260200160405190810160405280929190818152602001828054611b0390615d33565b8015611b505780601f10611b2557610100808354040283529160200191611b50565b820191906000526020600020905b815481529060010190602001808311611b3357829003601f168201915b5050505050905090565b6000600b80549050905090565b611b79611b72613065565b8383614353565b5050565b611b85612bfb565b600080600090505b84849050811015611c2157848482818110611bab57611baa615fa1565b5b905060200201359150611bbd82612c45565b611bd96008600084815260200190815260200160002084613745565b817ffec7db38481afeb8686a62ee7bba420143bd43540fe5e57b7316be50bdaa220c84604051611c099190615f1a565b60405180910390a2611c1a81613d59565b9050611b8d565b5050505050565b611c30612bfb565b81611c3a81612c45565b6000600860008581526020019081526020016000209050611c5a8161400e565b6000816000015490506000600b8281548110611c7957611c78615fa1565b5b9060005260206000209060030201905084816000015434611c9a9190616161565b14611cda576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611cd19061624d565b60405180910390fd5b611d46600a60008560030160009054906101000a900460a01b73ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000848152602001908152602001600020546144bf565b600a60008560030160009054906101000a900460a01b73ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600084815260200190815260200160002081905550611dbc838683600101546144cc565b857f1d9c4d2b3e13eb9ac08a42625750ac17ec6ca94b4755c49285e9467b4e48c89d86604051611dec9190615047565b60405180910390a2505050505050565b611e0d611e07613065565b836133b7565b611e4c576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611e4390615efa565b60405180910390fd5b611e58848484846145c5565b50505050565b611e66612bfb565b60008060005b85859050811015611f6757858582818110611e8a57611e89615fa1565b5b905060200201359250611e9c83612c45565b6008600084815260200190815260200160002091506000611ec08360020154612cbe565b14611f00576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611ef790615ce4565b60405180910390fd5b611f0983612d37565b611f138285612e85565b8373ffffffffffffffffffffffffffffffffffffffff16837fd964a27d45f595739c13d8b1160b57491050cacf3a2e5602207277d6228f64ee60405160405180910390a3611f6081613d59565b9050611e6c565b505050505050565b611f77612bfb565b60018383905011611fbd576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401611fb4906162b9565b60405180910390fd5b60003490506000806000808787905090505b60008111156122df57611fe1816144bf565b9050878782818110611ff657611ff5615fa1565b5b90506020020135935061200884612c45565b60086000858152602001908152602001600020925061202683613126565b60008360000154905060008460030160009054906101000a900460a01b9050600b828154811061205957612058615fa1565b5b9060005260206000209060030201935083600101548810156120b0576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016120a790616325565b60405180910390fd5b8360000154876120c09190616161565b96506120cf8560010154614621565b1561218a5761212f600960008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000848152602001908152602001600020546144bf565b600960008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff1916815260200190815260200160002060008481526020019081526020016000208190555061223c565b6121e5600a60008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000848152602001908152602001600020546144bf565b600a60008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000848152602001908152602001600020819055505b600083146122525761224d86612d37565b6122d8565b7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff856001018190555061228685888a6144cc565b89896040516122969291906163ba565b60405180910390207fb3f4c8ca702dbbd32d9a25ce17b1942a5060284d9d69fc4fcac8fb0397891b12888a6040516122cf9291906163d3565b60405180910390a25b5050611fcf565b5050505050505050565b60606122f48261301a565b60006122fe61464e565b9050600081511161231e5760405180602001604052806000815250612349565b8061232884614665565b604051602001612339929190616438565b6040516020818303038152906040525b915050919050565b612359613c78565b6000820361239c576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401612393906164a8565b60405180910390fd5b6000600c600084815260200190815260200160002060008381526020019081526020016000205414612403576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016123fa90616514565b60405180910390fd5b600b60405180606001604052808481526020018381526020014381525090806001815401808255809150506001900390600052602060002090600302016000909190919091506000820151816000015560208201518160010155604082015181600201555050600b80549050600c60008481526020019081526020016000206000838152602001908152602001600020819055507f6b39e3267efcd6611c8d7d2534c4715dcb4824322b90d85540a3a82967b6e7b782826040516124c89291906163d3565b60405180910390a15050565b60006124de612bfb565b6000821180156124f857503482866124f69190616534565b145b612537576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161252e906161e1565b60405180910390fd5b60006125438686613d66565b905061254e81613e17565b600160075401915060005b838110156125c25761256b8286613e62565b80836125779190616161565b7f1f44b78b04f7c6f80fc97ae8b196d9e9d7a81663114744f18fe5d073cd70ce4a8689896040516125aa93929190615ffd565b60405180910390a26125bb81613d59565b9050612559565b5050949350505050565b7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff81565b6125f8612bfb565b8161260281612c45565b60006008600085815260200190815260200160002090506126228161400e565b6000816000015490506000600b828154811061264157612640615fa1565b5b9060005260206000209060030201905080600101548511612697576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161268e9061624d565b60405180910390fd5b612703600a60008560030160009054906101000a900460a01b73ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000848152602001908152602001600020546144bf565b600a60008560030160009054906101000a900460a01b73ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600084815260200190815260200160002081905550612779838260000154876144cc565b857fc599168ac63ff28ec278088a2c424383a36ca26c931eb41af05e014f19252ea4866040516127a99190615047565b60405180910390a2505050505050565b60006127c3612bfb565b348251856127d19190616534565b14612811576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401612808906161e1565b60405180910390fd5b600061281d8585613d66565b905061282881613e17565b600160075401915060005b83518110156128d1576128608285838151811061285357612852615fa1565b5b6020026020010151613e62565b808361286c9190616161565b7f1f44b78b04f7c6f80fc97ae8b196d9e9d7a81663114744f18fe5d073cd70ce4a8583815181106128a05761289f615fa1565b5b602002602001015188886040516128b993929190615ffd565b60405180910390a26128ca81613d59565b9050612833565b50509392505050565b6128e2613c78565b43600b6128ef8484613d66565b81548110612900576128ff615fa1565b5b9060005260206000209060030201600201819055507f6b39e3267efcd6611c8d7d2534c4715dcb4824322b90d85540a3a82967b6e7b782826040516129469291906163d3565b60405180910390a15050565b61295a612bfb565b60008060005b84849050811015612a3f5784848281811061297e5761297d615fa1565b5b90506020020135925061299083612c45565b6008600084815260200190815260200160002091506129ae82613126565b60006129b983613ae7565b146129f9576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016129f090615f81565b60405180910390fd5b612a0282613b91565b827f11725367022c3ff288940f4b5473aa61c2da6a24af7363a1128ee2401e8983b260405160405180910390a2612a3881613d59565b9050612960565b5050505050565b6000600560008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060009054906101000a900460ff16905092915050565b612ae2613c78565b7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff600b612b0f8484613d66565b81548110612b2057612b1f615fa1565b5b9060005260206000209060030201600201819055507f099df2bf9247b43481cf1b791a4dd5fa1220c40c62940da539082fbcb30241d68282604051612b669291906163d3565b60405180910390a15050565b61ca8081565b612b80613c78565b600073ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff1603612bef576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401612be6906165e8565b60405180910390fd5b612bf88161422a565b50565b612c03611413565b15612c43576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401612c3a90616654565b60405180910390fd5b565b612c4e816115f6565b73ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff1614612cbb576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401612cb2906166c0565b60405180910390fd5b50565b6000612cc982614621565b612d08576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401612cff9061672c565b60405180910390fd5b600061ca8083612d189190616161565b9050438111612d2b576000915050612d32565b4381039150505b919050565b6000612d42826115f6565b9050612d52816000846001614733565b612d5b826115f6565b90506004600083815260200190815260200160002060006101000a81549073ffffffffffffffffffffffffffffffffffffffff02191690556001600360008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825403925050819055506002600083815260200190815260200160002060006101000a81549073ffffffffffffffffffffffffffffffffffffffff021916905581600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef60405160405180910390a4612e8181600084600161481e565b5050565b6000600b836000015481548110612e9f57612e9e615fa1565b5b906000526020600020906003020160000154905060008273ffffffffffffffffffffffffffffffffffffffff1682604051612ed99061677d565b60006040518083038185875af1925050503d8060008114612f16576040519150601f19603f3d011682016040523d82523d6000602084013e612f1b565b606091505b5050905080612f5f576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401612f56906167de565b60405180910390fd5b50505050565b60007f01ffc9a7000000000000000000000000000000000000000000000000000000007bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916827bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916149050919050565b612fd881614824565b613017576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161300e90616080565b60405180910390fd5b50565b61302381614824565b613062576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161305990616080565b60405180910390fd5b50565b600033905090565b816004600083815260200190815260200160002060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550808273ffffffffffffffffffffffffffffffffffffffff166130e0836115f6565b73ffffffffffffffffffffffffffffffffffffffff167f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b92560405160405180910390a45050565b6131338160020154614621565b15613173576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161316a9061684a565b60405180910390fd5b50565b60008260000154905060008360030160009054906101000a900460a01b905061319e84613ae7565b8310156131e0576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016131d790616325565b60405180910390fd5b6000613211600b84815481106131f9576131f8615fa1565b5b90600052602060002090600302016000015485613d66565b905061321c81613e17565b7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff85600101819055506132a0600960008473ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000858152602001908152602001600020546144bf565b600960008473ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff1916815260200190815260200160002060008581526020019081526020016000208190555080856000018190555061335a600a60008473ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600083815260200190815260200160002054613d59565b600a60008473ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000838152602001908152602001600020819055505050505050565b6000806133c3836115f6565b90508073ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff16148061340557506134048185612a46565b5b8061344357508373ffffffffffffffffffffffffffffffffffffffff1661342b84610d9c565b73ffffffffffffffffffffffffffffffffffffffff16145b91505092915050565b8273ffffffffffffffffffffffffffffffffffffffff1661346c826115f6565b73ffffffffffffffffffffffffffffffffffffffff16146134c2576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016134b9906168dc565b60405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff1603613531576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016135289061696e565b60405180910390fd5b61353e8383836001614733565b8273ffffffffffffffffffffffffffffffffffffffff1661355e826115f6565b73ffffffffffffffffffffffffffffffffffffffff16146135b4576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016135ab906168dc565b60405180910390fd5b6004600082815260200190815260200160002060006101000a81549073ffffffffffffffffffffffffffffffffffffffff02191690556001600360008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825403925050819055506001600360008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008282540192505081905550816002600083815260200190815260200160002060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550808273ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef60405160405180910390a4613740838383600161481e565b505050565b61374e82613126565b60008260000154905060008360030160009054906101000a900460a01b90508273ffffffffffffffffffffffffffffffffffffffff19168173ffffffffffffffffffffffffffffffffffffffff1916036137dd576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016137d49061624d565b60405180910390fd5b6137ea8460010154614621565b156139565761384a600960008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000848152602001908152602001600020546144bf565b600960008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000848152602001908152602001600020819055506138fb600960008573ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600084815260200190815260200160002054613d59565b600960008573ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600084815260200190815260200160002081905550613ab9565b6139b1600a60008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000848152602001908152602001600020546144bf565b600a60008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600084815260200190815260200160002081905550613a62600a60008573ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600084815260200190815260200160002054613d59565b600a60008573ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000848152602001908152602001600020819055505b828460030160006101000a8154816bffffffffffffffffffffffff021916908360a01c021790555050505050565b60008082600101549050613afa81614621565b613b39576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401613b30906169da565b60405180910390fd5b6000600b846000015481548110613b5357613b52615fa1565b5b90600052602060002090600302016001015482613b709190616161565b9050438111613b8457600092505050613b8c565b438103925050505b919050565b438160020181905550613c0a600960008360030160009054906101000a900460a01b73ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600083600001548152602001908152602001600020546144bf565b600960008360030160009054906101000a900460a01b73ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000836000015481526020019081526020016000208190555050565b613c80613065565b73ffffffffffffffffffffffffffffffffffffffff16613c9e611a62565b73ffffffffffffffffffffffffffffffffffffffff1614613cf4576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401613ceb90616a46565b60405180910390fd5b565b613cfe614865565b6000600660146101000a81548160ff0219169083151502179055507f5db9ee0a495bf2e6ff9c91a7834c1ba4fdd244a5e8aa4e537bd38aeae4b073aa613d42613065565b604051613d4f9190615135565b60405180910390a1565b6000600182019050919050565b600080600c6000858152602001908152602001600020600084815260200190815260200160002054905060008111613dd3576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401613dca90616ab2565b60405180910390fd5b613ddc816144bf565b91505092915050565b600043600b8381548110613dfc57613dfb615fa1565b5b90600052602060002090600302016002015411159050919050565b613e2081613de5565b613e5f576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401613e5690616b1e565b60405180910390fd5b50565b613e6d600754613d59565b60078190555060405180608001604052808381526020017fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff81526020017fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff81526020018273ffffffffffffffffffffffffffffffffffffffff191681525060086000600754815260200190815260200160002060008201518160000155602082015181600101556040820151816002015560608201518160030160006101000a8154816bffffffffffffffffffffffff021916908360a01c0217905550905050613fa8600a60008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600084815260200190815260200160002054613d59565b600a60008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff1916815260200190815260200160002060008481526020019081526020016000208190555061400a336007546148ae565b5050565b61401b8160010154614621565b1561405b576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161405290616b8a565b60405180910390fd5b50565b60008160000154905060008260030160009054906101000a900460a01b90504383600101819055506140e1600a60008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff191681526020019081526020016000206000848152602001908152602001600020546144bf565b600a60008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600084815260200190815260200160002081905550614192600960008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600084815260200190815260200160002054613d59565b600960008373ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600084815260200190815260200160002081905550505050565b60006002600083815260200190815260200160002060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff169050919050565b6000600660009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905081600660006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055508173ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff167f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e060405160405180910390a35050565b6142f8612bfb565b6001600660146101000a81548160ff0219169083151502179055507f62e78cea01bee320cd4e420270b5ea74000d11b0c9f74754ebdbfc544b05a25861433c613065565b6040516143499190615135565b60405180910390a1565b8173ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff16036143c1576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016143b890616bf6565b60405180910390fd5b80600560008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060006101000a81548160ff0219169083151502179055508173ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff167f17307eab39ab6107e8899845ad3d59bd9653f200f220920489ca2b5937696c31836040516144b29190614ff0565b60405180910390a3505050565b6000600182039050919050565b60006144d88383613d66565b90506144e381613e17565b61454f600a60008660030160009054906101000a900460a01b73ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff19168152602001908152602001600020600083815260200190815260200160002054613d59565b600a60008660030160009054906101000a900460a01b73ffffffffffffffffffffffffffffffffffffffff191673ffffffffffffffffffffffffffffffffffffffff1916815260200190815260200160002060008381526020019081526020016000208190555080846000018190555050505050565b6145d084848461344c565b6145dc848484846148cc565b61461b576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161461290616c88565b60405180910390fd5b50505050565b60007fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff8214159050919050565b606060405180602001604052806000815250905090565b60606000600161467484614a53565b01905060008167ffffffffffffffff811115614693576146926158b1565b5b6040519080825280601f01601f1916602001820160405280156146c55781602001600182028036833780820191505090505b509050600082602001820190505b600115614728578080600190039150507f3031323334353637383961626364656600000000000000000000000000000000600a86061a8153600a858161471c5761471b616ca8565b5b049450600085036146d3575b819350505050919050565b60018114614776576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161476d90616d23565b60405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff1614806147cd57506147cb6008600084815260200190815260200160002060020154614621565b155b61480c576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161480390616d8f565b60405180910390fd5b61481884848484614ba6565b50505050565b50505050565b60008073ffffffffffffffffffffffffffffffffffffffff16614846836141ed565b73ffffffffffffffffffffffffffffffffffffffff1614159050919050565b61486d611413565b6148ac576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016148a390616dfb565b60405180910390fd5b565b6148c8828260405180602001604052806000815250614bac565b5050565b60006148ed8473ffffffffffffffffffffffffffffffffffffffff16614c07565b15614a46578373ffffffffffffffffffffffffffffffffffffffff1663150b7a02614916613065565b8786866040518563ffffffff1660e01b81526004016149389493929190616e70565b6020604051808303816000875af192505050801561497457506040513d601f19601f820116820180604052508101906149719190616ed1565b60015b6149f6573d80600081146149a4576040519150601f19603f3d011682016040523d82523d6000602084013e6149a9565b606091505b5060008151036149ee576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016149e590616c88565b60405180910390fd5b805181602001fd5b63150b7a0260e01b7bffffffffffffffffffffffffffffffffffffffffffffffffffffffff1916817bffffffffffffffffffffffffffffffffffffffffffffffffffffffff191614915050614a4b565b600190505b949350505050565b600080600090507a184f03e93ff9f4daa797ed6e38ed64bf6a1f0100000000000000008310614ab1577a184f03e93ff9f4daa797ed6e38ed64bf6a1f0100000000000000008381614aa757614aa6616ca8565b5b0492506040810190505b6d04ee2d6d415b85acef81000000008310614aee576d04ee2d6d415b85acef81000000008381614ae457614ae3616ca8565b5b0492506020810190505b662386f26fc100008310614b1d57662386f26fc100008381614b1357614b12616ca8565b5b0492506010810190505b6305f5e1008310614b46576305f5e1008381614b3c57614b3b616ca8565b5b0492506008810190505b6127108310614b6b576127108381614b6157614b60616ca8565b5b0492506004810190505b60648310614b8e5760648381614b8457614b83616ca8565b5b0492506002810190505b600a8310614b9d576001810190505b80915050919050565b50505050565b614bb68383614c2a565b614bc360008484846148cc565b614c02576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401614bf990616c88565b60405180910390fd5b505050565b6000808273ffffffffffffffffffffffffffffffffffffffff163b119050919050565b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff1603614c99576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401614c9090616f4a565b60405180910390fd5b614ca281614824565b15614ce2576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401614cd990616fb6565b60405180910390fd5b614cf0600083836001614733565b614cf981614824565b15614d39576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401614d3090616fb6565b60405180910390fd5b6001600360008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008282540192505081905550816002600083815260200190815260200160002060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550808273ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef60405160405180910390a4614e4360008383600161481e565b5050565b60405180606001604052806000815260200160008152602001600081525090565b6000604051905090565b600080fd5b600080fd5b6000819050919050565b614e8f81614e7c565b8114614e9a57600080fd5b50565b600081359050614eac81614e86565b92915050565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b6000614edd82614eb2565b9050919050565b614eed81614ed2565b8114614ef857600080fd5b50565b600081359050614f0a81614ee4565b92915050565b60008060408385031215614f2757614f26614e72565b5b6000614f3585828601614e9d565b9250506020614f4685828601614efb565b9150509250929050565b60007fffffffff0000000000000000000000000000000000000000000000000000000082169050919050565b614f8581614f50565b8114614f9057600080fd5b50565b600081359050614fa281614f7c565b92915050565b600060208284031215614fbe57614fbd614e72565b5b6000614fcc84828501614f93565b91505092915050565b60008115159050919050565b614fea81614fd5565b82525050565b60006020820190506150056000830184614fe1565b92915050565b60006020828403121561502157615020614e72565b5b600061502f84828501614e9d565b91505092915050565b61504181614e7c565b82525050565b600060208201905061505c6000830184615038565b92915050565b600081519050919050565b600082825260208201905092915050565b60005b8381101561509c578082015181840152602081019050615081565b60008484015250505050565b6000601f19601f8301169050919050565b60006150c482615062565b6150ce818561506d565b93506150de81856020860161507e565b6150e7816150a8565b840191505092915050565b6000602082019050818103600083015261510c81846150b9565b905092915050565b600061511f82614eb2565b9050919050565b61512f81615114565b82525050565b600060208201905061514a6000830184615126565b92915050565b61515981615114565b811461516457600080fd5b50565b60008135905061517681615150565b92915050565b6000806040838503121561519357615192614e72565b5b60006151a185828601615167565b92505060206151b285828601614e9d565b9150509250929050565b600080604083850312156151d3576151d2614e72565b5b60006151e185828601614e9d565b92505060206151f285828601614e9d565b9150509250929050565b60008060006060848603121561521557615214614e72565b5b600061522386828701615167565b935050602061523486828701615167565b925050604061524586828701614e9d565b9150509250925092565b60007fffffffffffffffffffffffff000000000000000000000000000000000000000082169050919050565b6152848161524f565b811461528f57600080fd5b50565b6000813590506152a18161527b565b92915050565b600080604083850312156152be576152bd614e72565b5b60006152cc85828601614e9d565b92505060206152dd85828601615292565b9150509250929050565b600080fd5b600080fd5b600080fd5b60008083601f84011261530c5761530b6152e7565b5b8235905067ffffffffffffffff811115615329576153286152ec565b5b602083019150836020820283011115615345576153446152f1565b5b9250929050565b6000806020838503121561536357615362614e72565b5b600083013567ffffffffffffffff81111561538157615380614e77565b5b61538d858286016152f6565b92509250509250929050565b600081519050919050565b600082825260208201905092915050565b6000819050602082019050919050565b600081519050919050565b600082825260208201905092915050565b6000819050602082019050919050565b6153fa81614e7c565b82525050565b600061540c83836153f1565b60208301905092915050565b6000602082019050919050565b6000615430826153c5565b61543a81856153d0565b9350615445836153e1565b8060005b8381101561547657815161545d8882615400565b975061546883615418565b925050600181019050615449565b5085935050505092915050565b600061548f8383615425565b905092915050565b6000602082019050919050565b60006154af82615399565b6154b981856153a4565b9350836020820285016154cb856153b5565b8060005b8581101561550757848403895281516154e88582615483565b94506154f383615497565b925060208a019950506001810190506154cf565b50829750879550505050505092915050565b6000602082019050818103600083015261553381846154a4565b905092915050565b6155448161524f565b82525050565b600060a08201905061555f6000830188615038565b61556c6020830187615038565b6155796040830186615038565b6155866060830185615038565b615593608083018461553b565b9695505050505050565b60008083601f8401126155b3576155b26152e7565b5b8235905067ffffffffffffffff8111156155d0576155cf6152ec565b5b6020830191508360208202830111156155ec576155eb6152f1565b5b9250929050565b60008060006040848603121561560c5761560b614e72565b5b600084013567ffffffffffffffff81111561562a57615629614e77565b5b6156368682870161559d565b9350935050602061564986828701614e9d565b9150509250925092565b6000806020838503121561566a57615669614e72565b5b600083013567ffffffffffffffff81111561568857615687614e77565b5b6156948582860161559d565b92509250509250929050565b6000602082840312156156b6576156b5614e72565b5b60006156c484828501615167565b91505092915050565b600081519050919050565b600082825260208201905092915050565b6000819050602082019050919050565b60608201600082015161570f60008501826153f1565b50602082015161572260208501826153f1565b50604082015161573560408501826153f1565b50505050565b600061574783836156f9565b60608301905092915050565b6000602082019050919050565b600061576b826156cd565b61577581856156d8565b9350615780836156e9565b8060005b838110156157b1578151615798888261573b565b97506157a383615753565b925050600181019050615784565b5085935050505092915050565b600060208201905081810360008301526157d88184615760565b905092915050565b6157e981614fd5565b81146157f457600080fd5b50565b600081359050615806816157e0565b92915050565b6000806040838503121561582357615822614e72565b5b600061583185828601615167565b9250506020615842858286016157f7565b9150509250929050565b60008060006040848603121561586557615864614e72565b5b600084013567ffffffffffffffff81111561588357615882614e77565b5b61588f8682870161559d565b935093505060206158a286828701615292565b9150509250925092565b600080fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b6158e9826150a8565b810181811067ffffffffffffffff82111715615908576159076158b1565b5b80604052505050565b600061591b614e68565b905061592782826158e0565b919050565b600067ffffffffffffffff821115615947576159466158b1565b5b615950826150a8565b9050602081019050919050565b82818337600083830152505050565b600061597f61597a8461592c565b615911565b90508281526020810184848401111561599b5761599a6158ac565b5b6159a684828561595d565b509392505050565b600082601f8301126159c3576159c26152e7565b5b81356159d384826020860161596c565b91505092915050565b600080600080608085870312156159f6576159f5614e72565b5b6000615a0487828801615167565b9450506020615a1587828801615167565b9350506040615a2687828801614e9d565b925050606085013567ffffffffffffffff811115615a4757615a46614e77565b5b615a53878288016159ae565b91505092959194509250565b600080600060408486031215615a7857615a77614e72565b5b600084013567ffffffffffffffff811115615a9657615a95614e77565b5b615aa28682870161559d565b93509350506020615ab586828701614efb565b9150509250925092565b60008060008060808587031215615ad957615ad8614e72565b5b6000615ae787828801614e9d565b9450506020615af887828801614e9d565b9350506040615b0987828801615292565b9250506060615b1a87828801614e9d565b91505092959194509250565b600067ffffffffffffffff821115615b4157615b406158b1565b5b602082029050602081019050919050565b6000615b65615b6084615b26565b615911565b90508083825260208201905060208402830185811115615b8857615b876152f1565b5b835b81811015615bb15780615b9d8882615292565b845260208401935050602081019050615b8a565b5050509392505050565b600082601f830112615bd057615bcf6152e7565b5b8135615be0848260208601615b52565b91505092915050565b600080600060608486031215615c0257615c01614e72565b5b6000615c1086828701614e9d565b9350506020615c2186828701614e9d565b925050604084013567ffffffffffffffff811115615c4257615c41614e77565b5b615c4e86828701615bbb565b9150509250925092565b60008060408385031215615c6f57615c6e614e72565b5b6000615c7d85828601615167565b9250506020615c8e85828601615167565b9150509250929050565b7f6e6f7420726561647920746f2077697468647261770000000000000000000000600082015250565b6000615cce60158361506d565b9150615cd982615c98565b602082019050919050565b60006020820190508181036000830152615cfd81615cc1565b9050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b60006002820490506001821680615d4b57607f821691505b602082108103615d5e57615d5d615d04565b5b50919050565b7f4552433732313a20617070726f76616c20746f2063757272656e74206f776e6560008201527f7200000000000000000000000000000000000000000000000000000000000000602082015250565b6000615dc060218361506d565b9150615dcb82615d64565b604082019050919050565b60006020820190508181036000830152615def81615db3565b9050919050565b7f4552433732313a20617070726f76652063616c6c6572206973206e6f7420746f60008201527f6b656e206f776e6572206f7220617070726f76656420666f7220616c6c000000602082015250565b6000615e52603d8361506d565b9150615e5d82615df6565b604082019050919050565b60006020820190508181036000830152615e8181615e45565b9050919050565b7f4552433732313a2063616c6c6572206973206e6f7420746f6b656e206f776e6560008201527f72206f7220617070726f76656400000000000000000000000000000000000000602082015250565b6000615ee4602d8361506d565b9150615eef82615e88565b604082019050919050565b60006020820190508181036000830152615f1381615ed7565b9050919050565b6000602082019050615f2f600083018461553b565b92915050565b7f6e6f7420726561647920746f20756e7374616b65000000000000000000000000600082015250565b6000615f6b60148361506d565b9150615f7682615f35565b602082019050919050565b60006020820190508181036000830152615f9a81615f5e565b9050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052603260045260246000fd5b600060208284031215615fe657615fe5614e72565b5b6000615ff484828501615292565b91505092915050565b6000606082019050616012600083018661553b565b61601f6020830185615038565b61602c6040830184615038565b949350505050565b7f4552433732313a20696e76616c696420746f6b656e2049440000000000000000600082015250565b600061606a60188361506d565b915061607582616034565b602082019050919050565b600060208201905081810360008301526160998161605d565b9050919050565b7f4552433732313a2061646472657373207a65726f206973206e6f74206120766160008201527f6c6964206f776e65720000000000000000000000000000000000000000000000602082015250565b60006160fc60298361506d565b9150616107826160a0565b604082019050919050565b6000602082019050818103600083015261612b816160ef565b9050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b600061616c82614e7c565b915061617783614e7c565b925082820190508082111561618f5761618e616132565b5b92915050565b7f696e76616c696420706172616d65746572730000000000000000000000000000600082015250565b60006161cb60128361506d565b91506161d682616195565b602082019050919050565b600060208201905081810360008301526161fa816161be565b9050919050565b7f696e76616c6964206f7065726174696f6e000000000000000000000000000000600082015250565b600061623760118361506d565b915061624282616201565b602082019050919050565b600060208201905081810360008301526162668161622a565b9050919050565b7f696e76616c6964206c656e677468000000000000000000000000000000000000600082015250565b60006162a3600e8361506d565b91506162ae8261626d565b602082019050919050565b600060208201905081810360008301526162d281616296565b9050919050565b7f696e76616c6964206475726174696f6e00000000000000000000000000000000600082015250565b600061630f60108361506d565b915061631a826162d9565b602082019050919050565b6000602082019050818103600083015261633e81616302565b9050919050565b600081905092915050565b600080fd5b82818337505050565b600061636a8385616345565b93507f07ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff83111561639d5761639c616350565b5b6020830292506163ae838584616355565b82840190509392505050565b60006163c782848661635e565b91508190509392505050565b60006040820190506163e86000830185615038565b6163f56020830184615038565b9392505050565b600081905092915050565b600061641282615062565b61641c81856163fc565b935061642c81856020860161507e565b80840191505092915050565b60006164448285616407565b91506164508284616407565b91508190509392505050565b7f616d6f756e7420697320696e76616c6964000000000000000000000000000000600082015250565b600061649260118361506d565b915061649d8261645c565b602082019050919050565b600060208201905081810360008301526164c181616485565b9050919050565b7f6475706c6963617465206275636b657420747970650000000000000000000000600082015250565b60006164fe60158361506d565b9150616509826164c8565b602082019050919050565b6000602082019050818103600083015261652d816164f1565b9050919050565b600061653f82614e7c565b915061654a83614e7c565b925082820261655881614e7c565b9150828204841483151761656f5761656e616132565b5b5092915050565b7f4f776e61626c653a206e6577206f776e657220697320746865207a65726f206160008201527f6464726573730000000000000000000000000000000000000000000000000000602082015250565b60006165d260268361506d565b91506165dd82616576565b604082019050919050565b60006020820190508181036000830152616601816165c5565b9050919050565b7f5061757361626c653a2070617573656400000000000000000000000000000000600082015250565b600061663e60108361506d565b915061664982616608565b602082019050919050565b6000602082019050818103600083015261666d81616631565b9050919050565b7f6e6f74206f776e65720000000000000000000000000000000000000000000000600082015250565b60006166aa60098361506d565b91506166b582616674565b602082019050919050565b600060208201905081810360008301526166d98161669d565b9050919050565b7f6e6f7420616e20756e7374616b6564206275636b657400000000000000000000600082015250565b600061671660168361506d565b9150616721826166e0565b602082019050919050565b6000602082019050818103600083015261674581616709565b9050919050565b600081905092915050565b50565b600061676760008361674c565b915061677282616757565b600082019050919050565b60006167888261675a565b9150819050919050565b7f6661696c656420746f207472616e736665720000000000000000000000000000600082015250565b60006167c860128361506d565b91506167d382616792565b602082019050919050565b600060208201905081810360008301526167f7816167bb565b9050919050565b7f6e6f742061207374616b656420746f6b656e0000000000000000000000000000600082015250565b600061683460128361506d565b915061683f826167fe565b602082019050919050565b6000602082019050818103600083015261686381616827565b9050919050565b7f4552433732313a207472616e736665722066726f6d20696e636f72726563742060008201527f6f776e6572000000000000000000000000000000000000000000000000000000602082015250565b60006168c660258361506d565b91506168d18261686a565b604082019050919050565b600060208201905081810360008301526168f5816168b9565b9050919050565b7f4552433732313a207472616e7366657220746f20746865207a65726f2061646460008201527f7265737300000000000000000000000000000000000000000000000000000000602082015250565b600061695860248361506d565b9150616963826168fc565b604082019050919050565b600060208201905081810360008301526169878161694b565b9050919050565b7f6e6f7420616e20756e6c6f636b6564206275636b657400000000000000000000600082015250565b60006169c460168361506d565b91506169cf8261698e565b602082019050919050565b600060208201905081810360008301526169f3816169b7565b9050919050565b7f4f776e61626c653a2063616c6c6572206973206e6f7420746865206f776e6572600082015250565b6000616a3060208361506d565b9150616a3b826169fa565b602082019050919050565b60006020820190508181036000830152616a5f81616a23565b9050919050565b7f696e76616c6964206275636b6574207479706500000000000000000000000000600082015250565b6000616a9c60138361506d565b9150616aa782616a66565b602082019050919050565b60006020820190508181036000830152616acb81616a8f565b9050919050565b7f696e616374697665206275636b65742074797065000000000000000000000000600082015250565b6000616b0860148361506d565b9150616b1382616ad2565b602082019050919050565b60006020820190508181036000830152616b3781616afb565b9050919050565b7f6e6f742061206c6f636b656420746f6b656e0000000000000000000000000000600082015250565b6000616b7460128361506d565b9150616b7f82616b3e565b602082019050919050565b60006020820190508181036000830152616ba381616b67565b9050919050565b7f4552433732313a20617070726f766520746f2063616c6c657200000000000000600082015250565b6000616be060198361506d565b9150616beb82616baa565b602082019050919050565b60006020820190508181036000830152616c0f81616bd3565b9050919050565b7f4552433732313a207472616e7366657220746f206e6f6e20455243373231526560008201527f63656976657220696d706c656d656e7465720000000000000000000000000000602082015250565b6000616c7260328361506d565b9150616c7d82616c16565b604082019050919050565b60006020820190508181036000830152616ca181616c65565b9050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601260045260246000fd5b7f6261746368207472616e73666572206973206e6f7420737570706f7274656400600082015250565b6000616d0d601f8361506d565b9150616d1882616cd7565b602082019050919050565b60006020820190508181036000830152616d3c81616d00565b9050919050565b7f63616e6e6f74207472616e7366657220756e7374616b656420746f6b656e0000600082015250565b6000616d79601e8361506d565b9150616d8482616d43565b602082019050919050565b60006020820190508181036000830152616da881616d6c565b9050919050565b7f5061757361626c653a206e6f7420706175736564000000000000000000000000600082015250565b6000616de560148361506d565b9150616df082616daf565b602082019050919050565b60006020820190508181036000830152616e1481616dd8565b9050919050565b600081519050919050565b600082825260208201905092915050565b6000616e4282616e1b565b616e4c8185616e26565b9350616e5c81856020860161507e565b616e65816150a8565b840191505092915050565b6000608082019050616e856000830187615126565b616e926020830186615126565b616e9f6040830185615038565b8181036060830152616eb18184616e37565b905095945050505050565b600081519050616ecb81614f7c565b92915050565b600060208284031215616ee757616ee6614e72565b5b6000616ef584828501616ebc565b91505092915050565b7f4552433732313a206d696e7420746f20746865207a65726f2061646472657373600082015250565b6000616f3460208361506d565b9150616f3f82616efe565b602082019050919050565b60006020820190508181036000830152616f6381616f27565b9050919050565b7f4552433732313a20746f6b656e20616c7265616479206d696e74656400000000600082015250565b6000616fa0601c8361506d565b9150616fab82616f6a565b602082019050919050565b60006020820190508181036000830152616fcf81616f93565b905091905056fea2646970667358221220d4fb1d4a5f31b44bb4783b1e7268c7deb8646c92668ae8c0866d990b17ca737864736f6c63430008120033`
+)
+
+func TestLiquidStaking(t *testing.T) {
+	require := require.New(t)
+
+	testReadContract := func(cfg config.Config, t *testing.T) {
+		ctx := context.Background()
+
+		// Create a new blockchain
+		svr, err := itx.NewServer(cfg)
+		require.NoError(err)
+		require.NoError(svr.Start(ctx))
+		defer func() {
+			require.NoError(svr.Stop(ctx))
+		}()
+
+		chainID := cfg.Chain.ID
+		bc := svr.ChainService(chainID).Blockchain()
+		sf := svr.ChainService(chainID).StateFactory()
+		ap := svr.ChainService(chainID).ActionPool()
+		dao := svr.ChainService(chainID).BlockDAO()
+		registry := svr.ChainService(chainID).Registry()
+		require.NotNil(bc)
+		require.NotNil(registry)
+		admin := identityset.PrivateKey(26)
+		state0 := hash.BytesToHash160(identityset.Address(26).Bytes())
+		s := &state.Account{}
+		_, err = sf.State(s, protocol.LegacyKeyOption(state0))
+		require.NoError(err)
+		require.Equal(unit.ConvertIotxToRau(100000000), s.Balance)
+
+		// deploy staking contract
+		data, err := hex.DecodeString(_liquidStakingContractByteCode)
+		require.NoError(err)
+		fixedTime := time.Unix(cfg.Genesis.Timestamp, 0)
+		ex, err := action.SignedExecution(action.EmptyAddress, admin, 1, big.NewInt(0), 10000000, big.NewInt(testutil.TestGasPriceInt64), data)
+		require.NoError(err)
+
+		deployHash, err := ex.Hash()
+		require.NoError(err)
+		require.NoError(ap.Add(context.Background(), ex))
+		blk, err := bc.MintNewBlock(fixedTime)
+		require.NoError(err)
+		require.NoError(bc.CommitBlock(blk))
+		r, err := dao.GetReceiptByActionHash(deployHash, 1)
+		require.NoError(err)
+		require.Equal(r.ContractAddress, "io123vqxxup8n3ld8jygvx729r6295pv9krjn2tjh")
+
+		// set value
+		// contractABI, err := abi.JSON(strings.NewReader(blockindex.LiquidStakingContractABI))
+		// require.NoError(err)
+		fixedAmount := unit.ConvertIotxToRau(200)
+		// sk := identityset.PrivateKey(26)
+		nonce := uint64(0)
+		// data, err = contractABI.Pack("addBucketType", big.NewInt(10), big.NewInt(10000))
+		// require.NoError(err)
+		// require.True(len(data) > 0)
+		data, err = hex.DecodeString(`c8e7792300000000000000000000000000000000000000000000000000000000000003e80000000000000000000000000000000000000000000000000000000000002710`)
+		require.NoError(err)
+		ex, err = action.SignedExecution(r.ContractAddress, admin, nonce+2, fixedAmount, 10000000, big.NewInt(testutil.TestGasPriceInt64), data)
+		require.NoError(err)
+		require.NoError(ap.Add(context.Background(), ex))
+		// data, err = contractABI.Pack("get")
+		// require.NoError(err)
+		// require.True(len(data) > 0)
+		// ex, err = action.SignedExecution(r.ContractAddress, sk, nonce+2, fixedAmount, 1000000, big.NewInt(testutil.TestGasPriceInt64), data)
+		// require.NoError(err)
+		// require.NoError(ap.Add(context.Background(), ex))
+		blk, err = bc.MintNewBlock(fixedTime)
+		require.NoError(err)
+		require.NoError(bc.CommitBlock(blk))
+	}
+
+	cfg := config.Default
+	testTriePath, err := testutil.PathOfTempFile("trie")
+	require.NoError(err)
+	testDBPath, err := testutil.PathOfTempFile("db")
+	require.NoError(err)
+	testIndexPath, err := testutil.PathOfTempFile("index")
+	require.NoError(err)
+	testBloomfilterIndexPath, err := testutil.PathOfTempFile("bloomfilterindex")
+	require.NoError(err)
+	testCandidateIndexPath, err := testutil.PathOfTempFile("candidateindex")
+	require.NoError(err)
+	testLiquidStakeIndexPath, err := testutil.PathOfTempFile("liquidstakeindex")
+	require.NoError(err)
+	testSystemLogPath, err := testutil.PathOfTempFile("systemlog")
+	require.NoError(err)
+	testConsensusPath, err := testutil.PathOfTempFile("consensus")
+	require.NoError(err)
+	defer func() {
+		testutil.CleanupPath(testTriePath)
+		testutil.CleanupPath(testDBPath)
+		testutil.CleanupPath(testIndexPath)
+		testutil.CleanupPath(testBloomfilterIndexPath)
+		testutil.CleanupPath(testCandidateIndexPath)
+		testutil.CleanupPath(testLiquidStakeIndexPath)
+		testutil.CleanupPath(testSystemLogPath)
+		testutil.CleanupPath(testConsensusPath)
+		// clear the gateway
+		delete(cfg.Plugins, config.GatewayPlugin)
+	}()
+
+	cfg.ActPool.MinGasPriceStr = "0"
+	cfg.Chain.TrieDBPatchFile = ""
+	cfg.Chain.TrieDBPath = testTriePath
+	cfg.Chain.ChainDBPath = testDBPath
+	cfg.Chain.IndexDBPath = testIndexPath
+	cfg.Chain.BloomfilterIndexDBPath = testBloomfilterIndexPath
+	cfg.Chain.CandidateIndexDBPath = testCandidateIndexPath
+	cfg.Chain.LiquidStakingIndexDBPath = testLiquidStakeIndexPath
+	cfg.System.SystemLogDBPath = testSystemLogPath
+	cfg.Consensus.RollDPoS.ConsensusDBPath = testConsensusPath
+	cfg.Chain.ProducerPrivKey = "a000000000000000000000000000000000000000000000000000000000000000"
+	cfg.Consensus.Scheme = config.RollDPoSScheme
+	cfg.Genesis.NumDelegates = 1
+	cfg.Genesis.NumSubEpochs = 10
+	cfg.Genesis.Delegates = []genesis.Delegate{
+		{
+			OperatorAddrStr: identityset.Address(0).String(),
+			RewardAddrStr:   identityset.Address(0).String(),
+			VotesStr:        "10",
+		},
+	}
+	cfg.Genesis.PollMode = "lifeLong"
+	cfg.Genesis.EnableGravityChainVoting = false
+	cfg.Genesis.ActionGasLimit = 100000000
+	cfg.Plugins[config.GatewayPlugin] = true
+	cfg.Chain.EnableAsyncIndexWrite = false
+	cfg.Genesis.AleutianBlockHeight = 0
+	cfg.Genesis.BeringBlockHeight = 0
+	cfg.Genesis.HawaiiBlockHeight = 0
+
+	t.Run("test read staking contract", func(t *testing.T) {
+		testReadContract(cfg, t)
+	})
+}
